consolidate NativeTypes
