+++ src/SEDManagerCLI/PBA.cpp
@@ -3,7 +3,7 @@
 #include "Utility.hpp"
 #include <async++/join.hpp>
 
-#include <Archive/Types/ValueToNative.hpp>
+#include <Messaging/Native.hpp>
 #include <StorageDevice/StorageDevice.hpp>
 
 #include <EncryptedDevice/EncryptedDevice.hpp>
