+++ src/lib/MockDevice/State.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Messaging/NativeTypes.hpp>
+#include <Messaging/Native.hpp>
 #include <Messaging/Value.hpp>
 #include <TrustedPeripheral/MethodUtils.hpp>
