+++ test/Messaging/TestMethod.cpp
@@ -1,6 +1,6 @@
 #include <async++/join.hpp>
 
-#include <Archive/Types/ValueToNative.hpp>
+#include <Messaging/Native.hpp>
 #include <Messaging/Method.hpp>
 #include <Specification/Core/Defs/UIDs.hpp>
