+++ test/Messaging/TestUID.cpp
@@ -1,4 +1,4 @@
-#include <Messaging/NativeTypes.hpp>
+#include <Messaging/Native.hpp>
 
 #include <catch2/catch_test_macros.hpp>
